fix: 控制中心-声音-扬声器模块，扬声器输出音量为0时显示1e-6%；
功能与主线SP2保持一致；

Log: 取整数显示；
Task: 控制中心sp2迁移KLU，功能以主线为准；

+++ src/frame/window/modules/sound/speakerpage.cpp
@@ -155,7 +155,7 @@ void SpeakerPage::initSlider()
         m_speakSlider->blockSignals(true);
         m_speakSlider->setValue(v * 100 + 0.000001);
         m_speakSlider->blockSignals(false);
-        m_outputSlider->setValueLiteral(QString::number(v * 100 + 0.000001) + "%");
+        m_outputSlider->setValueLiteral(QString::number(static_cast<int>(v * 100 + 0.000001)) + "%");
     });
 
     connect(m_model, &SoundModel::maxUIVolumeChanged, this, [ = ](double maxvalue) {
@@ -174,7 +174,7 @@ void SpeakerPage::initSlider()
         m_speakSlider->blockSignals(true);
         m_speakSlider->setValue(static_cast<int>(m_model->speakerVolume() * PercentageNum + 0.000001));
         m_speakSlider->blockSignals(false);
-        m_outputSlider->setValueLiteral(QString::number(m_model->speakerVolume() * PercentageNum + 0.000001) + "%");
+        m_outputSlider->setValueLiteral(QString::number(static_cast<int>(m_model->speakerVolume() * PercentageNum + 0.000001)) + "%");
     });
 
     m_layout->insertWidget(1, m_outputSlider);
